Updated PiPano.py

+++ PiPano.py
@@ -12,7 +12,7 @@ degreesPerStep = 7.5*4
 
 # total number of steps
 numSteps = totalRangeInDegrees / degreesPerStep
-halfSteps = math.ceil(numSteps/2)
+halfSteps = int(math.ceil(numSteps/2))
 print '#steps: ' + str(numSteps)
 print 'halfSteps: ' + str(halfSteps)
